requirements.txt
# Swimming Pool Detection System - Dependencies
# Python 3.8+ required

# Core Dependencies
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
albumentations>=1.3.0

# Dataset APIs
kaggle>=1.5.0
roboflow>=1.0.0
python-dotenv>=1.0.0

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0
